requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
requests==2.22.0
djangorestframework==3.10.3